requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.7
Flask==1.1.1
Flask-Login==0.4.1
future==0.17.1
idna==2.8
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
pymongo==3.9.0
pytz==2019.3
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.7.7
Werkzeug==0.15.6
